Custom Split Implemented
Lots of changes:

I changed the SelectRoommateActivity to only display checkboxes for
roommates other than current user. The current user will always be
involved with the bill being made, so I eliminated the chance of a user
forgetting to select themselves. I changed the EvenSplit code that Tom
and Mandy wrote to work with this change.

CustomSplit works now. A user enters the amounts that his roommates owe
to him, and it gets stored in Parse.

I deleted some files we were no longer using as well.