Sean, the Underdog Story: Episode 1 - "Sean fixes EvenSplit"
I un-broke the even split. There is a problem with numList (and it’s
not because my code broke it). If you check a box, and then change your
mind and uncheck it, a 0 is included in numList still. So our app
breaks if anyone changes their mind on SelectRoommateActivity.